Implement bead services for generating random and multiple random integers.

+++ src/main/java/ru/rkhamatyarov/convivialatmosphere/service/RandomBeadGeneratorService.java
@@ -1,11 +1,9 @@
 package ru.rkhamatyarov.convivialatmosphere.service;
 
-import java.math.BigInteger;
-
 public interface RandomBeadGeneratorService {
 
     /**
      * @return return bead with randomly generated radius
      */
-    BigInteger generateRandomBeadFactor();
+    Integer generateRandomBeadFactor();
 }
